// dx/frontend/src/api/model/extractionStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * dx API
 * OpenAPI spec version: 0.1.0
 */

export type ExtractionStatus =
  (typeof ExtractionStatus)[keyof typeof ExtractionStatus];

export const ExtractionStatus = {
  pending: "pending",
  running: "running",
  succeeded: "succeeded",
  partial: "partial",
  failed: "failed",
} as const;