// dx/frontend/src/api/model/searchHitOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * dx API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentNodeOut } from "./documentNodeOut";

export interface SearchHitOut {
  document_id: string;
  node: DocumentNodeOut | null;
  offset: number;
  snippet: string;
  title: string;
}